+++ src/lib/commandclass/MultilevelSensorCC.ts
@@ -306,7 +306,7 @@ interface MultilevelSensorCCGetSpecificOptions {
 }
 type MultilevelSensorCCGetOptions =
 	| CCCommandOptions
-	| CCCommandOptions & MultilevelSensorCCGetSpecificOptions;
+	| (CCCommandOptions & MultilevelSensorCCGetSpecificOptions);
 
 @CCCommand(MultilevelSensorCommand.Get)
 @expectedCCResponse(MultilevelSensorCCReport)