+++ src/lib/commandclass/MultilevelSwitchCC.test.ts
@@ -1,0 +1,223 @@
+import { createEmptyMockDriver } from "../../../test/mocks";
+import { IDriver } from "../driver/IDriver";
+import { Duration } from "../values/Duration";
+import { CommandClasses } from "./CommandClasses";
+import {
+	MultilevelSwitchCC,
+	MultilevelSwitchCCGet,
+	MultilevelSwitchCCReport,
+	MultilevelSwitchCCSet,
+	MultilevelSwitchCCStartLevelChange,
+	MultilevelSwitchCCStopLevelChange,
+	MultilevelSwitchCCSupportedGet,
+	MultilevelSwitchCommand,
+} from "./MultilevelSwitchCC";
+
+const fakeDriver = (createEmptyMockDriver() as unknown) as IDriver;
+
+function buildCCBuffer(nodeId: number, payload: Buffer): Buffer {
+	return Buffer.concat([
+		Buffer.from([
+			nodeId, // node number
+			payload.length + 1, // remaining length
+			CommandClasses["Multilevel Switch"], // CC
+		]),
+		payload,
+	]);
+}
+
+describe("lib/commandclass/MultilevelSwitchCC => ", () => {
+	it("the Get command should serialize correctly", () => {
+		const cc = new MultilevelSwitchCCGet(fakeDriver, { nodeId: 1 });
+		const expected = buildCCBuffer(
+			1,
+			Buffer.from([
+				MultilevelSwitchCommand.Get, // CC Command
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the Set command (V1) should serialize correctly", () => {
+		const cc = new MultilevelSwitchCCSet(fakeDriver, {
+			nodeId: 2,
+			targetValue: 55,
+		});
+		const expected = buildCCBuffer(
+			2,
+			Buffer.from([
+				MultilevelSwitchCommand.Set, // CC Command
+				55, // target value
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the Set command (V2) should serialize correctly", () => {
+		const cc = new MultilevelSwitchCCSet(fakeDriver, {
+			nodeId: 2,
+			targetValue: 55,
+			duration: new Duration(2, "minutes"),
+		});
+		const expected = buildCCBuffer(
+			2,
+			Buffer.from([
+				MultilevelSwitchCommand.Set, // CC Command
+				55, // target value,
+				0x81, // 2 minutes
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the Report command (V1) should be deserialized correctly", () => {
+		const ccData = buildCCBuffer(
+			1,
+			Buffer.from([
+				MultilevelSwitchCommand.Report, // CC Command
+				55, // current value
+			]),
+		);
+		const cc = new MultilevelSwitchCCReport(fakeDriver, { data: ccData });
+
+		expect(cc.currentValue).toBe(55);
+		expect(cc.targetValue).toBeUndefined();
+		expect(cc.duration).toBeUndefined();
+	});
+
+	it("the Report command (v4) should be deserialized correctly", () => {
+		const ccData = buildCCBuffer(
+			1,
+			Buffer.from([
+				MultilevelSwitchCommand.Report, // CC Command
+				55, // current value
+				66, // target value
+				1, // duration
+			]),
+		);
+		const cc = new MultilevelSwitchCCReport(fakeDriver, { data: ccData });
+
+		expect(cc.currentValue).toBe(55);
+		expect(cc.targetValue).toBe(66);
+		expect(cc.duration!.unit).toBe("seconds");
+		expect(cc.duration!.value).toBe(1);
+	});
+
+	it("the StartLevelChange command (V1) should serialize correctly (up, ignore start level)", () => {
+		const cc = new MultilevelSwitchCCStartLevelChange(fakeDriver, {
+			nodeId: 2,
+			direction: "up",
+			ignoreStartLevel: true,
+		});
+		const expected = buildCCBuffer(
+			2,
+			Buffer.from([
+				MultilevelSwitchCommand.StartLevelChange, // CC Command
+				0b001_00000, // up, ignore start level,
+				0, // don't include a start level that should be ignored
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the StartLevelChange command (V1) should serialize correctly (down)", () => {
+		const cc = new MultilevelSwitchCCStartLevelChange(fakeDriver, {
+			nodeId: 2,
+			direction: "down",
+			ignoreStartLevel: false,
+			startLevel: 50,
+		});
+		const expected = buildCCBuffer(
+			2,
+			Buffer.from([
+				MultilevelSwitchCommand.StartLevelChange, // CC Command
+				0b010_00000, // down,
+				50,
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the StopLevelChange command should serialize correctly", () => {
+		const cc = new MultilevelSwitchCCStopLevelChange(fakeDriver, {
+			nodeId: 1,
+		});
+		const expected = buildCCBuffer(
+			1,
+			Buffer.from([
+				MultilevelSwitchCommand.StopLevelChange, // CC Command
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the StartLevelChange command (V2) should serialize correctly (down, with duration)", () => {
+		const cc = new MultilevelSwitchCCStartLevelChange(fakeDriver, {
+			nodeId: 2,
+			direction: "down",
+			ignoreStartLevel: false,
+			startLevel: 50,
+			duration: new Duration(3, "seconds"),
+		});
+		const expected = buildCCBuffer(
+			2,
+			Buffer.from([
+				MultilevelSwitchCommand.StartLevelChange, // CC Command
+				0b010_00000, // down,
+				50, // start level
+				3, // 3 sec
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("the SupportedGet command should serialize correctly", () => {
+		const cc = new MultilevelSwitchCCSupportedGet(fakeDriver, {
+			nodeId: 1,
+		});
+		const expected = buildCCBuffer(
+			1,
+			Buffer.from([
+				MultilevelSwitchCommand.SupportedGet, // CC Command
+			]),
+		);
+		expect(cc.serialize()).toEqual(expected);
+	});
+
+	it("deserializing an unsupported command should return an unspecified version of MultilevelSwitchCC", () => {
+		const serializedCC = buildCCBuffer(
+			1,
+			Buffer.from([255]), // not a valid command
+		);
+		const cc: any = new MultilevelSwitchCC(fakeDriver, {
+			data: serializedCC,
+		});
+		expect(cc.constructor).toBe(MultilevelSwitchCC);
+	});
+
+	// it("the CC values should have the correct metadata", () => {
+	// 	// Readonly, 0-99
+	// 	const currentValueMeta = getCCValueMetadata(
+	// 		CommandClasses.MultilevelSwitch,
+	// 		"currentValue",
+	// 	);
+	// 	expect(currentValueMeta).toMatchObject({
+	// 		readable: true,
+	// 		writeable: false,
+	// 		min: 0,
+	// 		max: 99,
+	// 	});
+
+	// 	// Writeable, 0-99
+	// 	const targetValueMeta = getCCValueMetadata(
+	// 		CommandClasses.MultilevelSwitch,
+	// 		"targetValue",
+	// 	);
+	// 	expect(targetValueMeta).toMatchObject({
+	// 		readable: true,
+	// 		writeable: true,
+	// 		min: 0,
+	// 		max: 99,
+	// 	});
+	// });
+});