+++ src/lib/driver/Driver.ts
@@ -40,7 +40,7 @@ import { getDefaultPriority, Message } from "../message/Message";
 import { InterviewStage, IZWaveNode, NodeStatus } from "../node/INode";
 import { isNodeQuery } from "../node/INodeQuery";
 import { ZWaveNode } from "../node/Node";
-import { DeepPartial, skipBytes } from "../util/misc";
+import { DeepPartial, getEnumMemberName, skipBytes } from "../util/misc";
 import { num2hex } from "../util/strings";
 import { DriverEventCallbacks, DriverEvents, IDriver } from "./IDriver";
 import { Transaction } from "./Transaction";
@@ -424,10 +424,18 @@ export class Driver extends EventEmitter implements IDriver {
 	): number {
 		if (this._controller == undefined || !this.controller.nodes.has(nodeId))
 			return 0;
-		return this.controller.nodes
+		const endpoint = this.controller.nodes
 			.get(nodeId)!
-			.getEndpoint(endpointIndex)!
-			.getCCVersion(cc);
+			.getEndpoint(endpointIndex);
+		if (!endpoint) {
+			throw new Error(
+				`getSupportedCCVersionForEndpoint failed. cc = ${getEnumMemberName(
+					CommandClasses,
+					cc,
+				)} nodeId = ${nodeId}, endpointIndex = ${endpointIndex}`,
+			);
+		}
+		return endpoint.getCCVersion(cc);
 	}
 
 	/**
@@ -527,7 +535,10 @@ export class Driver extends EventEmitter implements IDriver {
 		}
 	}
 
-	private _cleanupHandler = () => void this.destroy();
+	private _cleanupHandler = (): void => {
+		this.destroy();
+	};
+
 	/**
 	 * Terminates the driver instance and closes the underlying serial connection.
 	 * Must be called under any circumstances.
@@ -1039,7 +1050,7 @@ ${handlers.length} left`,
 
 		if (isNodeQuery(msg) || isCommandClassContainer(msg)) {
 			const node = msg.getNodeUnsafe();
-			if (node && node.status === NodeStatus.Dead) {
+			if (node?.status === NodeStatus.Dead) {
 				// We have received a message from a dead node, bring it back to life
 				// We do not know if the node is actually awake, so mark it as unknown for now
 				node.status = NodeStatus.Unknown;
@@ -1237,8 +1248,7 @@ ${handlers.length} left`,
 		this.currentTransaction = undefined;
 		// If a sleeping node has no messages pending, send it back to sleep
 		if (
-			node &&
-			node.supportsCC(CommandClasses["Wake Up"]) &&
+			node?.supportsCC(CommandClasses["Wake Up"]) &&
 			!this.hasPendingMessages(node)
 		) {
 			node.sendNoMoreInformation();
@@ -1297,7 +1307,7 @@ ${handlers.length} left`,
 		// Don't send messages to dead nodes
 		if (isNodeQuery(msg) || isCommandClassContainer(msg)) {
 			const node = msg.getNodeUnsafe();
-			if (node && node.status === NodeStatus.Dead) {
+			if (node?.status === NodeStatus.Dead) {
 				throw new ZWaveError(
 					"The message will not be sent because the node is presumed dead",
 					ZWaveErrorCodes.Controller_MessageDropped,
@@ -1502,10 +1512,7 @@ ${handlers.length} left`,
 		this.sortSendQueue();
 
 		// Don't forget the current transaction
-		if (
-			this.currentTransaction &&
-			this.currentTransaction.message.getNodeId() === nodeId
-		) {
+		if (this.currentTransaction?.message.getNodeId() === nodeId) {
 			// Change the priority to WakeUp and re-add it to the queue
 			this.currentTransaction.priority = MessagePriority.WakeUp;
 			this.sendQueue.add(this.currentTransaction);
@@ -1540,10 +1547,7 @@ ${handlers.length} left`,
 		this.sendQueue.remove(...transactionsToRemove);
 
 		// Don't forget the current transaction
-		if (
-			this.currentTransaction &&
-			this.currentTransaction.message.getNodeId() === nodeId
-		) {
+		if (this.currentTransaction?.message.getNodeId() === nodeId) {
 			this.rejectCurrentTransaction(
 				new ZWaveError(
 					errorMsg,