+++ src/lib/log/Driver.test.ts
@@ -58,7 +58,7 @@ function createTransaction(
 describe("lib/log/Driver =>", () => {
 	let driverLogger: winston.Logger;
 	let spyTransport: SpyTransport;
-	let wasSilenced = true;
+	const wasSilenced = true;
 
 	// Replace all defined transports with a spy transport
 	beforeAll(() => {